model/requirements.txt
torch>=2.0
yfinance>=0.2.28
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
